requirements.txt
openai
langgraph
python-dotenv
pydantic
typing-extensions